requirements.txt
Pillow
google-adk
google-genai